Move MempoolInterface to a separate module
